004 ServerCode/ImageClassificationAndOID/requirements.txt
Flask==3.1.0
tensorflow==2.18.0
Pillow==11.1.0